fix small problem
