+++ smartRGrailsPlugin.groovy
@@ -49,7 +49,7 @@ class smartRGrailsPlugin {
             clinicalDataResourceService = ref('clinicalDataResourceService')
             highDimExportService = ref('highDimExportService')
         }
-        dataQueryService(eae.plugin.EaeService) {
+        EaeDataService(eae.plugin.EaeDataService) {
             studiesResourceService = ref('studiesResourceService')
             conceptsResourceService = ref('conceptsResourceService')
             clinicalDataResourceService = ref('clinicalDataResourceService')
